Re-write multiserver.js and multiserver2.js to special sensors on A101 and k64f boards. Singed-off-by: Zhang Xiaoyu <xiaoyux.zhang@intel.com>
